+++ msal/src/test/java/com/microsoft/identity/client/robolectric/shadows/ShadowAuthority.java
@@ -1,4 +1,4 @@
-package com.microsoft.identity.client.shadows;
+package com.microsoft.identity.client.robolectric.shadows;
 
 import android.net.Uri;
