+++ msal/src/test/java/com/microsoft/identity/client/robolectric/tests/AcquireTokenNetworkTest.java
@@ -1,98 +1,59 @@
-package com.microsoft.identity.client.robolectric;
+package com.microsoft.identity.client.robolectric.tests;
 
 import android.app.Activity;
-import android.content.Context;
-import android.content.SharedPreferences;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.test.core.app.ApplicationProvider;
 
 import com.microsoft.identity.client.AcquireTokenParameters;
 import com.microsoft.identity.client.AcquireTokenSilentParameters;
 import com.microsoft.identity.client.AuthenticationCallback;
 import com.microsoft.identity.client.IAccount;
 import com.microsoft.identity.client.IAuthenticationResult;
-import com.microsoft.identity.client.IMultipleAccountPublicClientApplication;
 import com.microsoft.identity.client.IPublicClientApplication;
-import com.microsoft.identity.client.PublicClientApplication;
 import com.microsoft.identity.client.exception.MsalClientException;
 import com.microsoft.identity.client.exception.MsalException;
-import com.microsoft.identity.client.shadows.ShadowAuthority;
-import com.microsoft.identity.client.shadows.ShadowHttpRequest;
-import com.microsoft.identity.client.shadows.ShadowStorageHelper;
-import com.microsoft.identity.common.adal.internal.util.StringExtensions;
-import com.microsoft.identity.common.exception.ClientException;
-import com.microsoft.identity.common.internal.authorities.Authority;
-import com.microsoft.identity.common.internal.authorities.MockAuthority;
-import com.microsoft.identity.common.internal.cache.ICacheRecord;
-import com.microsoft.identity.common.internal.dto.CredentialType;
-import com.microsoft.identity.common.internal.providers.microsoft.microsoftsts.MicrosoftStsAuthorizationRequest;
-import com.microsoft.identity.common.internal.providers.oauth2.OAuth2Strategy;
-import com.microsoft.identity.common.internal.providers.oauth2.OAuth2TokenCache;
-import com.microsoft.identity.common.internal.providers.oauth2.TokenResponse;
-import com.microsoft.identity.common.internal.testutils.MockTokenResponse;
+import com.microsoft.identity.client.robolectric.shadows.ShadowAuthority;
+import com.microsoft.identity.client.robolectric.shadows.ShadowStorageHelper;
 import com.microsoft.identity.common.internal.util.StringUtil;
-import com.microsoft.identity.msal.R;
+import com.microsoft.identity.common.utilities.TestConfigurationHelper;
+import com.microsoft.identity.common.utilities.TestConfigurationQuery;
 
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mockito;
 import org.robolectric.RobolectricTestRunner;
-import org.robolectric.RuntimeEnvironment;
 import org.robolectric.annotation.Config;
-import org.robolectric.util.Scheduler;
 
 import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Locale;
-import java.util.Map;
-import java.util.Set;
 
-import static com.microsoft.identity.common.internal.cache.CacheKeyValueDelegate.CACHE_VALUE_SEPARATOR;
 import static junit.framework.Assert.fail;
 
 @RunWith(RobolectricTestRunner.class)
-@Config(shadows = {ShadowStorageHelper.class, ShadowAuthority.class, ShadowHttpRequest.class})
-public final class PublicClientApplicationMockTest {
+@Config(shadows = {ShadowStorageHelper.class, ShadowAuthority.class})
+public final class AcquireTokenNetworkTest {
 
     private static final String[] SCOPES = {"user.read", "openid", "offline_access", "profile"};
-    private static final String AAD_MOCK_AUTHORITY = "https://test.authority/aad.mock";
+    private static final String DEFAULT_AAD_AUTHORITY = "https://login.microsoftonline.com/common";
+    private static final String AAD_ROPC_TEST_AUTHORITY = "https://test.authority/aad.ropc";
 
     @Test
-    public void canGetAccount() {
-        final Context context = ApplicationProvider.getApplicationContext();
-
-        PublicClientApplication.create(context, R.raw.test_msal_config_multiple_account, new PublicClientApplication.ApplicationCreatedListener() {
-            @Override
-            public void onCreated(IPublicClientApplication application) {
-                final IAccount account = loadAccountForTest(application);
-                Assert.assertTrue(account != null);
-            }
-
-            @Override
-            public void onError(MsalException exception) {
-                fail(exception.getMessage());
-            }
-        });
-    }
-
-    @Test
-    public void canAcquireToken() {
-        new PublicClientApplicationBaseTest() {
+    public void canPerformROPC() {
+        new AcquireTokenBaseTest() {
 
             @Override
             void makeAcquireTokenCall(final IPublicClientApplication publicClientApplication,
                                       final Activity activity) {
 
-                String username = "fake@test.com";
+                TestConfigurationQuery query = new TestConfigurationQuery();
+                query.userType = "Member";
+                query.isFederated = false;
+                query.federationProvider = "ADFSv4";
+
+                String username = TestConfigurationHelper.getUpnForTest(query);
 
                 AcquireTokenParameters parameters = new AcquireTokenParameters.Builder()
                         .startAuthorizationFromActivity(activity)
                         .withLoginHint(username)
                         .withScopes(Arrays.asList(SCOPES))
-                        .fromAuthority(AAD_MOCK_AUTHORITY)
+                        .fromAuthority(AAD_ROPC_TEST_AUTHORITY)
                         .callback(new AuthenticationCallback() {
                             @Override
                             public void onSuccess(IAuthenticationResult authenticationResult) {
@@ -102,43 +63,6 @@ public final class PublicClientApplicationMockTest {
                             @Override
                             public void onError(MsalException exception) {
                                 fail(exception.getMessage());
-                            }
-
-                            @Override
-                            public void onCancel() {
-                                fail("User cancelled flow");
-                            }
-                        })
-                        .build();
-
-                publicClientApplication.acquireToken(parameters);
-                RoboTestUtils.flushScheduler();
-            }
-
-        }.performTest();
-    }
-
-    @Test
-    public void acquireTokenFailsIfLoginHintNotProvidedForRopc() {
-        new PublicClientApplicationBaseTest() {
-
-            @Override
-            void makeAcquireTokenCall(final IPublicClientApplication publicClientApplication,
-                                      final Activity activity) {
-
-                AcquireTokenParameters parameters = new AcquireTokenParameters.Builder()
-                        .startAuthorizationFromActivity(activity)
-                        .withScopes(Arrays.asList(SCOPES))
-                        .fromAuthority(AAD_MOCK_AUTHORITY)
-                        .callback(new AuthenticationCallback() {
-                            @Override
-                            public void onSuccess(IAuthenticationResult authenticationResult) {
-                                fail("Unexpected Success");
-                            }
-
-                            @Override
-                            public void onError(MsalException exception) {
-                                Assert.assertTrue(true);
                             }
 
                             @Override
@@ -157,18 +81,23 @@ public final class PublicClientApplicationMockTest {
 
     @Test
     public void canAcquireSilentAfterGettingToken() {
-        new PublicClientApplicationBaseTest() {
+        new AcquireTokenBaseTest() {
 
             @Override
             void makeAcquireTokenCall(final IPublicClientApplication publicClientApplication,
                                       final Activity activity) {
 
-                String username = "fake@test.com";
+                TestConfigurationQuery query = new TestConfigurationQuery();
+                query.userType = "Member";
+                query.isFederated = false;
+                query.federationProvider = "ADFSv4";
+
+                final String username = TestConfigurationHelper.getUpnForTest(query);
 
                 final AcquireTokenSilentParameters silentParameters = new AcquireTokenSilentParameters.Builder()
                         .withScopes(Arrays.asList(SCOPES))
                         .forceRefresh(false)
-                        .fromAuthority(AAD_MOCK_AUTHORITY)
+                        .fromAuthority(DEFAULT_AAD_AUTHORITY)
                         .callback(new AuthenticationCallback() {
                             @Override
                             public void onSuccess(IAuthenticationResult authenticationResult) {
@@ -191,7 +120,7 @@ public final class PublicClientApplicationMockTest {
                         .startAuthorizationFromActivity(activity)
                         .withLoginHint(username)
                         .withScopes(Arrays.asList(SCOPES))
-                        .fromAuthority(AAD_MOCK_AUTHORITY)
+                        .fromAuthority(AAD_ROPC_TEST_AUTHORITY)
                         .callback(new AuthenticationCallback() {
                             @Override
                             public void onSuccess(IAuthenticationResult authenticationResult) {
@@ -222,56 +151,24 @@ public final class PublicClientApplicationMockTest {
     }
 
     @Test
-    public void canAcquireSilentIfValidTokensAvailableInCache() {
-        new PublicClientApplicationBaseTest() {
-            @Override
-            void makeAcquireTokenCall(final IPublicClientApplication publicClientApplication,
-                                      final Activity activity) {
-                final IAccount account = loadAccountForTest(publicClientApplication);
-
-                final AcquireTokenSilentParameters silentParameters = new AcquireTokenSilentParameters.Builder()
-                        .withScopes(Arrays.asList(SCOPES))
-                        .forceRefresh(false)
-                        .fromAuthority(AAD_MOCK_AUTHORITY)
-                        .forAccount(account)
-                        .callback(new AuthenticationCallback() {
-                            @Override
-                            public void onSuccess(IAuthenticationResult authenticationResult) {
-                                Assert.assertTrue(!StringUtil.isEmpty(authenticationResult.getAccessToken()));
-                            }
-
-                            @Override
-                            public void onError(MsalException exception) {
-                                fail(exception.getMessage());
-                            }
-
-                            @Override
-                            public void onCancel() {
-                                fail("User cancelled flow");
-                            }
-                        })
-                        .build();
-
-                publicClientApplication.acquireTokenSilentAsync(silentParameters);
-                RoboTestUtils.flushScheduler();
-            }
-
-        }.performTest();
-    }
-
-    @Test
     public void forceRefreshWorks() {
-        new PublicClientApplicationBaseTest() {
+        new AcquireTokenBaseTest() {
 
             @Override
             void makeAcquireTokenCall(final IPublicClientApplication publicClientApplication,
                                       final Activity activity) {
-                final IAccount account = loadAccountForTest(publicClientApplication);
+
+                TestConfigurationQuery query = new TestConfigurationQuery();
+                query.userType = "Member";
+                query.isFederated = false;
+                query.federationProvider = "ADFSv4";
+
+                final String username = TestConfigurationHelper.getUpnForTest(query);
+
                 final AcquireTokenSilentParameters silentParameters = new AcquireTokenSilentParameters.Builder()
                         .withScopes(Arrays.asList(SCOPES))
                         .forceRefresh(true)
-                        .forAccount(account)
-                        .fromAuthority(AAD_MOCK_AUTHORITY)
+                        .fromAuthority(DEFAULT_AAD_AUTHORITY)
                         .callback(new AuthenticationCallback() {
                             @Override
                             public void onSuccess(IAuthenticationResult authenticationResult) {
@@ -290,34 +187,22 @@ public final class PublicClientApplicationMockTest {
                         })
                         .build();
 
-                publicClientApplication.acquireTokenSilentAsync(silentParameters);
-                RoboTestUtils.flushScheduler();
-            }
-
-        }.performTest();
-    }
-
-    @Test
-    public void silentCallFailsIfAccountNotProvided() {
-        new PublicClientApplicationBaseTest() {
-
-            @Override
-            void makeAcquireTokenCall(final IPublicClientApplication publicClientApplication,
-                                      final Activity activity) {
-
-                final AcquireTokenSilentParameters silentParameters = new AcquireTokenSilentParameters.Builder()
+                AcquireTokenParameters parameters = new AcquireTokenParameters.Builder()
+                        .startAuthorizationFromActivity(activity)
+                        .withLoginHint(username)
                         .withScopes(Arrays.asList(SCOPES))
-                        .forceRefresh(false)
-                        .fromAuthority(AAD_MOCK_AUTHORITY)
+                        .fromAuthority(AAD_ROPC_TEST_AUTHORITY)
                         .callback(new AuthenticationCallback() {
                             @Override
                             public void onSuccess(IAuthenticationResult authenticationResult) {
-                                fail("Unexpected success");
+                                Assert.assertTrue(!StringUtil.isEmpty(authenticationResult.getAccessToken()));
+                                IAccount account = authenticationResult.getAccount();
+                                silentParameters.setAccount(account);
                             }
 
                             @Override
                             public void onError(MsalException exception) {
-                                Assert.assertTrue(exception instanceof MsalClientException);
+                                fail(exception.getMessage());
                             }
 
                             @Override
@@ -327,6 +212,8 @@ public final class PublicClientApplicationMockTest {
                         })
                         .build();
 
+                publicClientApplication.acquireToken(parameters);
+                RoboTestUtils.flushScheduler();
                 publicClientApplication.acquireTokenSilentAsync(silentParameters);
                 RoboTestUtils.flushScheduler();
             }
@@ -336,19 +223,23 @@ public final class PublicClientApplicationMockTest {
 
     @Test
     public void silentCallFailsIfCacheIsEmpty() {
-        new PublicClientApplicationBaseTest() {
+        new AcquireTokenBaseTest() {
 
             @Override
             void makeAcquireTokenCall(final IPublicClientApplication publicClientApplication,
                                       final Activity activity) {
-                final IAccount account = loadAccountForTest(publicClientApplication);
-                RoboTestUtils.clearCache();
+
+                TestConfigurationQuery query = new TestConfigurationQuery();
+                query.userType = "Member";
+                query.isFederated = false;
+                query.federationProvider = "ADFSv4";
+
+                final String username = TestConfigurationHelper.getUpnForTest(query);
 
                 final AcquireTokenSilentParameters silentParameters = new AcquireTokenSilentParameters.Builder()
                         .withScopes(Arrays.asList(SCOPES))
                         .forceRefresh(false)
-                        .forAccount(account)
-                        .fromAuthority(AAD_MOCK_AUTHORITY)
+                        .fromAuthority(DEFAULT_AAD_AUTHORITY)
                         .callback(new AuthenticationCallback() {
                             @Override
                             public void onSuccess(IAuthenticationResult authenticationResult) {
@@ -367,6 +258,34 @@ public final class PublicClientApplicationMockTest {
                         })
                         .build();
 
+                AcquireTokenParameters parameters = new AcquireTokenParameters.Builder()
+                        .startAuthorizationFromActivity(activity)
+                        .withLoginHint(username)
+                        .withScopes(Arrays.asList(SCOPES))
+                        .fromAuthority(AAD_ROPC_TEST_AUTHORITY)
+                        .callback(new AuthenticationCallback() {
+                            @Override
+                            public void onSuccess(IAuthenticationResult authenticationResult) {
+                                Assert.assertTrue(!StringUtil.isEmpty(authenticationResult.getAccessToken()));
+                                IAccount account = authenticationResult.getAccount();
+                                silentParameters.setAccount(account);
+                            }
+
+                            @Override
+                            public void onError(MsalException exception) {
+                                fail(exception.getMessage());
+                            }
+
+                            @Override
+                            public void onCancel() {
+                                fail("User cancelled flow");
+                            }
+                        })
+                        .build();
+
+                publicClientApplication.acquireToken(parameters);
+                RoboTestUtils.flushScheduler();
+                RoboTestUtils.clearCache();
                 publicClientApplication.acquireTokenSilentAsync(silentParameters);
                 RoboTestUtils.flushScheduler();
             }
@@ -376,18 +295,23 @@ public final class PublicClientApplicationMockTest {
 
     @Test
     public void silentWorksWhenCacheHasNoAccessToken() {
-        new PublicClientApplicationBaseTest() {
+        new AcquireTokenBaseTest() {
 
             @Override
             void makeAcquireTokenCall(final IPublicClientApplication publicClientApplication,
                                       final Activity activity) {
-                final IAccount account = loadAccountForTest(publicClientApplication);
+
+                TestConfigurationQuery query = new TestConfigurationQuery();
+                query.userType = "Member";
+                query.isFederated = false;
+                query.federationProvider = "ADFSv4";
+
+                final String username = TestConfigurationHelper.getUpnForTest(query);
 
                 final AcquireTokenSilentParameters silentParameters = new AcquireTokenSilentParameters.Builder()
                         .withScopes(Arrays.asList(SCOPES))
                         .forceRefresh(false)
-                        .forAccount(account)
-                        .fromAuthority(AAD_MOCK_AUTHORITY)
+                        .fromAuthority(DEFAULT_AAD_AUTHORITY)
                         .callback(new AuthenticationCallback() {
                             @Override
                             public void onSuccess(IAuthenticationResult authenticationResult) {
@@ -406,6 +330,33 @@ public final class PublicClientApplicationMockTest {
                         })
                         .build();
 
+                AcquireTokenParameters parameters = new AcquireTokenParameters.Builder()
+                        .startAuthorizationFromActivity(activity)
+                        .withLoginHint(username)
+                        .withScopes(Arrays.asList(SCOPES))
+                        .fromAuthority(AAD_ROPC_TEST_AUTHORITY)
+                        .callback(new AuthenticationCallback() {
+                            @Override
+                            public void onSuccess(IAuthenticationResult authenticationResult) {
+                                Assert.assertTrue(!StringUtil.isEmpty(authenticationResult.getAccessToken()));
+                                IAccount account = authenticationResult.getAccount();
+                                silentParameters.setAccount(account);
+                            }
+
+                            @Override
+                            public void onError(MsalException exception) {
+                                fail(exception.getMessage());
+                            }
+
+                            @Override
+                            public void onCancel() {
+                                fail("User cancelled flow");
+                            }
+                        })
+                        .build();
+
+                publicClientApplication.acquireToken(parameters);
+                RoboTestUtils.flushScheduler();
                 RoboTestUtils.removeAccessTokenFromCache();
                 publicClientApplication.acquireTokenSilentAsync(silentParameters);
                 RoboTestUtils.flushScheduler();
@@ -414,114 +365,5 @@ public final class PublicClientApplicationMockTest {
         }.performTest();
     }
 
-    @Test
-    public void silentWorksWhenCacheHasExpiredAccessToken() {
-        new PublicClientApplicationBaseTest() {
-
-            @Override
-            void makeAcquireTokenCall(final IPublicClientApplication publicClientApplication,
-                                      final Activity activity) {
-                ICacheRecord cacheRecord = createDataInCacheWithExpiredAccessToken(publicClientApplication);
-                final String loginHint = cacheRecord.getAccount().getUsername();
-                final IAccount account = performGetAccount(publicClientApplication, loginHint);
-
-                final AcquireTokenSilentParameters silentParameters = new AcquireTokenSilentParameters.Builder()
-                        .withScopes(Arrays.asList(SCOPES))
-                        .forceRefresh(false)
-                        .forAccount(account)
-                        .fromAuthority(AAD_MOCK_AUTHORITY)
-                        .callback(new AuthenticationCallback() {
-                            @Override
-                            public void onSuccess(IAuthenticationResult authenticationResult) {
-                                Assert.assertTrue(!StringUtil.isEmpty(authenticationResult.getAccessToken()));
-                            }
-
-                            @Override
-                            public void onError(MsalException exception) {
-                                fail(exception.getMessage());
-                            }
-
-                            @Override
-                            public void onCancel() {
-                                fail("User cancelled flow");
-                            }
-                        })
-                        .build();
-                
-                publicClientApplication.acquireTokenSilentAsync(silentParameters);
-                RoboTestUtils.flushScheduler();
-            }
-
-        }.performTest();
-    }
-
-
-    private ICacheRecord saveTokens(TokenResponse tokenResponse, IPublicClientApplication application) throws ClientException {
-        OAuth2TokenCache tokenCache = application.getConfiguration().getOAuth2TokenCache();
-        String clientId = application.getConfiguration().getClientId();
-        Authority authority = new MockAuthority();
-        OAuth2Strategy strategy = authority.createOAuth2Strategy();
-        MicrosoftStsAuthorizationRequest fakeAuthRequest = Mockito.mock(MicrosoftStsAuthorizationRequest.class);
-        Mockito.when(fakeAuthRequest.getAuthority()).thenReturn(authority.getAuthorityURL());
-        Mockito.when(fakeAuthRequest.getClientId()).thenReturn(clientId);
-        return tokenCache.save(strategy, fakeAuthRequest, tokenResponse);
-    }
-
-    private IAccount performGetAccount(IPublicClientApplication application, final String loginHint) {
-        final IAccount[] requestedAccount = {null};
-        final IMultipleAccountPublicClientApplication multipleAcctApp = (IMultipleAccountPublicClientApplication) application;
-        multipleAcctApp.getAccount(
-                loginHint.trim(),
-                new IMultipleAccountPublicClientApplication.GetAccountCallback() {
-                    @Override
-                    public void onTaskCompleted(final IAccount account) {
-                        if (account != null) {
-                            requestedAccount[0] = account;
-                        } else {
-                            fail("No account found matching identifier");
-                        }
-                    }
-
-                    @Override
-                    public void onError(final MsalException exception) {
-                        fail("No account found matching identifier");
-                    }
-                });
-        RoboTestUtils.flushScheduler();
-        return requestedAccount[0];
-    }
-
-    private ICacheRecord createDataInCache(IPublicClientApplication application) {
-        ICacheRecord cacheRecord = null;
-        final TokenResponse tokenResponse = MockTokenResponse.getTokenResponse();
-
-        try {
-            cacheRecord = saveTokens(tokenResponse, application);
-        } catch (ClientException e) {
-            fail("Unable to save tokens to cache: " + e.getMessage());
-        }
-
-        return cacheRecord;
-    }
-
-    private ICacheRecord createDataInCacheWithExpiredAccessToken(IPublicClientApplication application) {
-        ICacheRecord cacheRecord = null;
-        final TokenResponse tokenResponse = MockTokenResponse.getTokenResponseWithExpiredAccessToken();
-
-        try {
-            cacheRecord = saveTokens(tokenResponse, application);
-        } catch (ClientException e) {
-            fail("Unable to save tokens to cache: " + e.getMessage());
-        }
-
-        return cacheRecord;
-    }
-
-    private IAccount loadAccountForTest(IPublicClientApplication application) {
-        ICacheRecord cacheRecord = createDataInCache(application);
-        final String loginHint = cacheRecord.getAccount().getUsername();
-        final IAccount account = performGetAccount(application, loginHint);
-        return account;
-    }
 
 }
