+++ msal/src/test/res/raw/test_config.json
